LayoutTests/http/tests/security/postMessage/origin-follows-lexical-scope.html

<!DOCTYPE html>
<html>
<head>
<script src="resources/recv.js"></script>
<script>
if (window.layoutTestController) {
    layoutTestController.dumpAsText();
    layoutTestController.waitUntilDone();
}

function listener(e) {
    var msg = 'Received message: data="' + e.data + '" origin="' + e.origin + '"';

    log(msg);

    if (e.data.match(/ready/)) {
        log("ready");
        frames[0].go();
    }

    if (e.data.match(/data="done"/) && window.layoutTestController)
        layoutTestController.notifyDone();
}

addEventListener("message", listener, false);

document.domain = "0.0.1";
</script>
<body>
<div>window.location.href = <script>document.write(window.location.href);</script></div>
<div>document.domain = <script>document.write(document.domain);</script></div>
<div id="result">waiting...</div>
<div><iframe src="http://127.0.0.1:8080/security/postMessage/resources/middle-frame-for-lexical.html" id="child" width="800" height="300"
  style="border: 1px solid black;">
</iframe></div>
</body>
</html>
